Fix Registration Client
