+++ src/App.vue
@@ -1,8 +1,8 @@
 <template>
   <Navbar />
   <!-- <Login/> -->
-  <RegPeopleBirja></RegPeopleBirja>
-  <!-- <RegClient></RegClient> -->
+  <!-- <RegPeopleBirja></RegPeopleBirja> -->
+  <RegClient></RegClient>
   <!-- <Users></Users> -->
   <!-- <Roles></Roles> -->
   <!-- <Access></Access> -->
