+++ src/assets/RegPeopleBirja.css
@@ -5,6 +5,7 @@
 
 .wrapper-stepper {
   background-color: #fff;
+
   padding: 60px;
   /* box-shadow: rgba(0, 0, 0, 0.09); */
 }
