+++ src/components/Registration/RegStepsClient/RegStepOne.vue
@@ -1,0 +1,39 @@
+<template>
+    <h1 class="text-2xl font-bold">Регистрация клиент РКП</h1>
+    <h1 class="text-lg mt-2 font-semibold text-black opacity-30 inline-block">Уже есть учетная запись?</h1>
+    <a href="#" class="ml-2 text-sky-600 text-lg mt-2 font-semibold opacity-100">Войти в систему</a>
+    <div class="shadow-lg p-4 lg:p-16 my-12">
+        <p>
+            Lorem ipsum, dolor sit amet consectetur adipisicing elit. Officia, iusto. Tempora accusamus
+            corrupti
+            dolore iure quibusdam neque, reiciendis eveniet labore ipsum exercitationem officia cumque
+            facere
+            blanditiis repudiandae sequi voluptatibus ex culpa temporibus distinctio assumenda quae cum quis
+            recusandae. Fugiat nobis aliquam placeat repellendus ducimus totam officia iusto blanditiis
+            quidem
+            tempore sed ab cum aut, culpa architecto autem expedita. Quaerat eos accusamus nostrum,
+            recusandae
+            beatae sit eligendi. Eaque reprehenderit numquam dolorem nemo soluta expedita, vero tempore, cum
+            natus accusamus illum, nam ducimus quod cumque maiores vitae rerum nobis. Id assumenda saepe
+            dolorem
+            sapiente quia iusto hic aut. Aspernatur doloremque impedit, maxime corrupti vel similique,
+            incidunt
+            modi ducimus exercitationem aut adipisci amet officia aliquam voluptatibus dignissimos quibusdam
+            nisi necessitatibus. Dolore error et praesentium soluta ducimus nemo, obcaecati illum dolorem
+            ratione facere unde reiciendis adipisci deserunt, excepturi odit tempore iusto! Ullam, ab
+            itaque.
+            Soluta dolorem accusantium laborum esse, facere cum quia illum, velit, recusandae iusto id quo?
+            Maxime, deserunt accusantium consequuntur suscipit praesentium totam repellat iusto aut.
+            Sapiente
+            assumenda fugiat harum iure consequuntur, asperiores alias magni culpa, quo quisquam, expedita
+            eveniet iusto labore enim accusantium! Facilis, libero velit. Mollitia asperiores id quae neque
+            sapiente nesciunt, nihil quasi vel exercitationem aspernatur atque, porro autem.
+        </p>
+    </div>
+    <div class="flex items-center mb-4">
+        <input id="default-checkbox" required type="checkbox" value=""
+            class="w-4 h-4 text-blue-600 bg-gray-100 border-gray-300 rounded focus:ring-blue-500 dark:focus:ring-blue-600 dark:ring-offset-gray-800 focus:ring-2 dark:bg-gray-700 dark:border-gray-600">
+        <label for="default-checkbox" class="ml-2 text-sm font-bold text-gray-900 dark:text-gray-300">Соглашаюсь с условиями
+            оферты</label>
+    </div>
+</template>
